refactor: move lynx setup to helpers.rs
